% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ChainPlots.R
\name{ChainPlots}
\alias{ChainPlots}
\title{Coefficient history trace plots from a Reversible Jump results object}
\usage{
ChainPlots(results, vars.to.include = NULL, var.dictionary = NULL,
  predictors.only = FALSE, file = NULL, add.to.title = NULL,
  par.mfrow = NULL, add.panel.labels = FALSE)
}
\arguments{
\item{results}{A \code{\link[R2BGLiMS]{R2BGLiMS_Results-class}} object frum running the \code{\link[R2BGLiMS]{R2BGLiMS}} command.}

\item{predictors.only}{If true only chain plots for predictors are plotted (and intercept,
likelihood etc. are excluded). Default is FALSE.}

\item{file}{PDF path}

\item{add.to.title}{Optional character string to add to the start of each plot title}

\item{par.mfrow}{Allow passing the mfrow layout vector to par. Default par(mfrow=c(5,2))}

\item{add.panel.labels}{Add lower case letter labels to panel titles. Default FALSE.
Note: Does not work when providing a var.dictionary}
}
\value{
NA
}
\description{
Generates a PDF of chain plots for all variables from a Reversible Jump results object
}
\examples{
library(R2BGLiMS)

### --- Logistic regression with two model space prior components
# Load results from the biopsy logistic regression analysis (see R2MHRJ examples)
data(biopsyResults)
ChainPlots(biopsyResults)
}
\author{
Paul Newcombe
}

