% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CredibleSet.R
\name{CredibleSet}
\alias{CredibleSet}
\title{Infer credible sets of predictors}
\usage{
CredibleSet(results, credible.percentile.threshold = 0.95)
}
\arguments{
\item{results}{A \code{\link[R2BGLiMS]{R2BGLiMS_Results-class}} object frum running the \code{\link[R2BGLiMS]{R2BGLiMS}} command.}

\item{credible.percentile.threshold}{Percentile to determine credible set for (default 0.95 for a 95\% Credible Set).}
}
\value{
List of variables included in the credible set.
}
\description{
Infer credible sets of predictors from a Reversible Jump Results object.
}
\author{
Paul Newcombe
}

