% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiddenFunctions.R
\name{.ApproxPostProbsByModelEnumeration}
\alias{.ApproxPostProbsByModelEnumeration}
\title{Enumertated approximate posterior probabilities}
\usage{
.ApproxPostProbsByModelEnumeration(enumerated.model.likelihood.scores,
  model.space.priors, enumerate.up.to.dim)
}
\arguments{
\item{enumerate.up.to.dim}{Maxmimum dimension to truncate space too (can be less than or equal to
what was passed to R2BGLiMS). If left NULL the value passed to R2BGLiMS will be used.}

\item{results}{R2BGLiMS results object}
}
\value{
A list containing all model, marginal and dimension approximate posterior probabilities
}
\description{
Calulcates approximate posterior probabiltiies from a Reversible Jump Results object
for an anlaysis which has been run with enumeration turned on for the beginning.
}
\author{
Paul Newcombe
}
