% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiddenFunctions.R
\name{.BayesFactor}
\alias{.BayesFactor}
\title{Calculates a Bayes Facotor, give a prior and posterior probabiltiies. Check2}
\usage{
.BayesFactor(prior.prob, post.prob)
}
\arguments{
\item{prior.prob}{prior probability}

\item{post.prob}{posterior probability}
}
\value{
Bayes Factor
}
\description{
Calculates a Bayes Facotor, give a prior and posterior probabiltiies. Check2
}
\author{
Paul Newcombe
}
