% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiddenFunctions.R
\name{.BetaBinomialProbabilitySpecificModel}
\alias{.BetaBinomialProbabilitySpecificModel}
\title{Calculates a Beta-Binomial prior probability for a SPECIFIC model. From Bottolo et al.
This is not the prior on a particular dimension (that would required the binomial co-efficient).}
\usage{
.BetaBinomialProbabilitySpecificModel(k, n, a, b)
}
\arguments{
\item{k}{dimension of specific model}

\item{n}{total number of covariates}

\item{a}{beta-binomial hyper parameter a}

\item{b}{beta-binomial hyper parameter b}
}
\value{
Probability
}
\description{
Calculates a Beta-Binomial prior probability for a SPECIFIC model. From Bottolo et al.
This is not the prior on a particular dimension (that would required the binomial co-efficient).
}
\author{
Paul Newcombe
}
