% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiddenFunctions.R
\name{.ConfounderAdjustedResiduals}
\alias{.ConfounderAdjustedResiduals}
\title{Replaces the outcome variable with residuals from a linear regression on the confounders. I.e. removes effect
of confounders for the conjugate models.}
\usage{
.ConfounderAdjustedResiduals(data, outcome.var, confounders)
}
\arguments{
\item{data}{matrix or data.frame containing outcome and confounders}

\item{outcome.var}{name of outcome.variable}

\item{confounders}{list of confounders}
}
\value{
data original dataset with the confounders removed and the outcome replaces with adjusted residuals
}
\description{
Replaces the outcome variable with residuals from a linear regression on the confounders. I.e. removes effect
of confounders for the conjugate models.
}
\author{
Paul Newcombe
}
