% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiddenFunctions.R
\name{.ModelSpaceProbs}
\alias{.ModelSpaceProbs}
\title{Calculates prior probababilities for x, and >=x causal variables, when a truncated Poisson prior is used for model space}
\usage{
.ModelSpaceProbs(V, mu)
}
\arguments{
\item{V}{total number of variables}

\item{mu}{model space mean}
}
\value{
prior matrix whereby 1st row is prior probabilities for =x causal variables, and 2nd row for >=x variables (x = 1,...V)
}
\description{
Calculates prior probababilities for x, and >=x causal variables, when a truncated Poisson prior is used for model space
}
\author{
Paul Newcombe
}
