% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiddenFunctions.R
\name{.ModelSpaceSpecProb}
\alias{.ModelSpaceSpecProb}
\title{Calculates prior probabability of causality for a particular variable, when a Poisson prior is used for model space}
\usage{
.ModelSpaceSpecProb(V, mu, n.var = 1)
}
\arguments{
\item{V}{total number of variables}

\item{mu}{model space mean}

\item{n.var}{number of specific variables (default is 1)}
}
\value{
prior probability for a single variable
}
\description{
Calculates prior probabability of causality for a particular variable, when a Poisson prior is used for model space
}
\author{
Paul Newcombe
}
