% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiddenFunctions.R
\name{.ReadData}
\alias{.ReadData}
\title{Read formatted data file}
\usage{
.ReadData(data.file)
}
\arguments{
\item{data.file}{data file to be read}
}
\value{
dataRead a list containing 
covariates: matrix of covariate values
disease: binary outcome vector
n: number of subjects
R: number of clusters
startRJ: 1st variable included in RJ 
V: number of variables 
var.names: names
}
\description{
Reads in a .txt data file, formatted for my Java program, and returns as a list
}
\author{
Paul Newcombe
}
