% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/JAM_PointEstimate.R
\name{JAM_PointEstimate}
\alias{JAM}
\alias{JAM_PointEstimate}
\title{JAM (Joint Analysis of Marginal statistics) multivariate point estimator.}
\usage{
JAM(marginal.betas = NULL, X.ref = NULL)
}
\arguments{
\item{marginal.betas}{Vector of marginal effect estimates to re-analyse with JAM under multivariate models.}

\item{X.ref}{Reference genotype matrix used by JAM to impute the SNP-SNP correlations. If multiple regions are to be
analysed this should be a list containing reference genotype matrices for each region. Individual's genotype
must be coded as a numeric risk allele count 0/1/2. Non-integer values reflecting imputation may be given.
NB: The risk allele coding MUST correspond to that used in marginal.betas. These matrices must each be positive definite and
the column names must correspond to the names of the marginal.betas vector.}
}
\value{
A vector of multivariate point estimates.

The function \code{\link{PrettyResultsTable}} can be used to print summary posterior results for all parameters. Other functions
for summarising results are listed under "see also".
}
\description{
Constructs a point estimate of multivariate SNP effects, based on the JAM estimator, from a vetorc of marginal
effects and a reference genotype matrix.
}
\author{
Paul Newcombe
}
\seealso{
See also \code{\link{JAM}}.
}

