% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/JAM_PointEstimates.R, R/JAM_PosteriorSampler.R
\name{JAM_PointEstimates}
\alias{JAM_PointEstimates}
\alias{JAM_PosteriorSampler}
\title{JAM (Joint Analysis of Marginal statistics) multivariate point estimator.}
\usage{
JAM_PointEstimates(marginal.betas = NULL, X.ref = NULL, n = NULL,
  just.get.z = FALSE)

JAM_PosteriorSampler(marginal.betas = NULL, X.ref = NULL)
}
\arguments{
\item{marginal.betas}{Vector of marginal effect estimates to re-analyse with JAM under multivariate models.}

\item{X.ref}{Reference genotype matrix used by JAM to impute the SNP-SNP correlations. If multiple regions are to be
analysed this should be a list containing reference genotype matrices for each region. Individual's genotype
must be coded as a numeric risk allele count 0/1/2. Non-integer values reflecting imputation may be given.
NB: The risk allele coding MUST correspond to that used in marginal.betas. These matrices must each be positive definite and
the column names must correspond to the names of the marginal.betas vector.}

\item{just.get.z}{Return the estimated X'y outcome used in JAM (without actually calculating the corresponding point estimates). This is mainly
for internal use.}
}
\value{
A vector of multivariate point estimates.

A vector of multivariate point estimates.
}
\description{
Constructs a point estimate of multivariate SNP effects, based on the JAM estimator, from a vetorc of marginal
effects and a reference genotype matrix.

Constructs a point estimate of multivariate SNP effects, based on the JAM estimator, from a vetorc of marginal
effects and a reference genotype matrix.
}
\author{
Paul Newcombe

Paul Newcombe
}
\seealso{
See also \code{\link{JAM}}.

See also \code{\link{JAM}}.
}

