% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JAM_RankCheck.R
\name{JAM_RankCheck}
\alias{JAM_RankCheck}
\title{JAM rank check.}
\usage{
JAM_RankCheck(X.ref = NULL)
}
\arguments{
\item{X.ref}{Reference genotype matrix used by JAM to impute the SNP-SNP correlations. If multiple regions are to be 
analysed this should be a list containing reference genotype matrices for each region. Individual's genotype
must be coded as a numeric risk allele count 0/1/2. Non-integer values reflecting imputation may be given.
NB: The risk allele coding MUST correspond to that used in marginal.betas. These matrices must each be positive definite and
the column names must correspond to the names of the marginal.betas vector.}
}
\value{
TRUE/FALSE of whether the matrix is full rank (hopefully it is and TRUE is returned). If X.ref is a list
of reference matrices for different regions, a vector of boolen TRUE/FALSE values is returned, one for each LD
block.
}
\description{
Checks the rank of a reference genotype matrix, which is required to run JAM.
}
\seealso{
See also \code{\link{JAM}}.
}
\author{
Paul Newcombe
}
