% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JAMMR_ManhattanPlot.R
\name{JAMMR_ManhattanPlot}
\alias{JAMMR_ManhattanPlot}
\title{JAMMR_ManhattanPlot}
\usage{
JAMMR_ManhattanPlot(jammr.results, only.best = FALSE)
}
\arguments{
\item{jammr.results}{A \code{\link[R2BGLiMS]{JAMMR}} results object frum running the \code{\link[R2BGLiMS]{JAMMR}} command.}

\item{only.best}{APOSTOLOS TO PROVIDE}
}
\description{
Manhattan plotting function for results from \code{\link[R2BGLiMS]{JAMMR}} analyses.
}
\author{
Apostolos Gkatzionis
}
