% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JAMMR_SensitivityPlot.R
\name{JAMMR_SensitivityPlot}
\alias{JAMMR_SensitivityPlot}
\title{JAMMR_SensitivityPlot}
\usage{
JAMMR_SensitivityPlot(
  jammr.results,
  adjust.se = TRUE,
  line.at.zero = TRUE,
  show.n.snps = FALSE
)
}
\arguments{
\item{jammr.results}{A \code{\link[R2BGLiMS]{JAMMR}} results object frum running the \code{\link[R2BGLiMS]{JAMMR}} command.}

\item{adjust.se}{APOSTOLOS TO PROVIDE}

\item{line.at.zero}{Whether to include a line at zero}

\item{show.n.snps}{APOSTOLOS TO PROVIDE}
}
\description{
Plots causal effect estimates and confidence intervals for multiple w runs
from \code{\link[R2BGLiMS]{JAMMR}} analyses.
}
\author{
Apostolos Gkatzionis
}
