% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/JamPosteriorSample.R
\name{JamPosteriorSample}
\alias{JamPosteriorSample}
\title{Generate JAM posterior sample}
\usage{
JamPosteriorSample(xTx = NULL, z = NULL, sigma2_invGamma_a = NULL,
  sigma2_invGamma_b = NULL, g.prior = TRUE, tau = NULL, model = NULL,
  n.samples = 1000)
}
\arguments{
\item{xTx}{GaussianMarg and GaussianMargConj ONLY: List containing each block's plug-in estimate for X'X.}

\item{z}{GaussianMarg and GaussianMargConj ONLY: Vector of quantities calculated from the summary statistics.}

\item{sigma2_invGamma_a}{Guassian models ONLY: Inverse-Gamma parameter one for the residual
precision. For the conjugate model this parameter is intergrated out, and this may be provided as in the
Bottolo and Richardson 2010 notation. For an informative prior for the non-conjugate model
(taking into account Java parameterisation) choose N/2.}

\item{sigma2_invGamma_b}{Guassian models ONLY: Inverse-Gamma parameter one for the residual
precision. For the conjugate model this parameter is intergrated out, and this may be provided as in the
Bottolo and Richardson 2010 notation. For an informative prior for the non-conjugate model
(taking into account Java parameterisation) choose N/(2*variance estimate).}

\item{g.prior}{Gaussian conjugate models ONLY: Whether to use a g-prior for the beta's - i.e. a multivariate normal
with correlation structure proportional to sigma^2*X'X^-1 or to use independence priors (default = FALSE).}

\item{tau}{Gaussian conjugate models ONLY: Value to use for sparsity parameter tau (tau*sigma^2 parameterisation).
Default residual.var.n. If modelling this parameter, this value is used to center the Zellner-Siow prior
and as an intial value.}

\item{model}{Character vector of covariate names to include in the model.}

\item{n.samples}{Number of posterior samples.}

\item{likelihood}{Type of model to fit. Current options are "Logistic" (for binary data), "Weibull" (for survival data),
"Gaussian" (for continuous data), "GaussianConj" (linear regression exploiting conjugate results), "GaussianMarg" (for analysis of univariate associations from Gaussian linear
regressions) and "GaussianMargConj" (for analysis under a marginal conjugate linear regression model).}
}
\value{
The posterior sample as a matrix. Rows are different posterior samples, and columns are the covairates
specified in model.
}
\description{
Conjugate results are used to generate a posterior sample of parameters (beta and sigma) under the JAM framework,
conditional on a particular model, i.e. selection of SNPs.
}
\author{
Paul Newcombe
}

