% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JAMPred_ParallelBlockIndices.R
\name{JAMPred_ParallelBlockIndices}
\alias{JAMPred_ParallelBlockIndices}
\title{List of which LD blocks are to be analysed on which CPU core}
\usage{
JAMPred_ParallelBlockIndices(n.cores, n.blocks.to.analyse)
}
\arguments{
\item{n.cores}{Number of CPU cores to use}

\item{n.blocks.to.analyse}{Total number of LD blocks to analyse}
}
\value{
A list of which blocks are to be analysed on each CPU core.
}
\description{
For a given number of CPU cores and LD blocks, this function returns a list 
indicating which blocks should be analysed on each CPU core.
}
\author{
Paul Newcombe
}
