% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JAMPred_Step1ParallelAnalyses.R
\name{JAMPred_Step1ParallelAnalyses}
\alias{JAMPred_Step1ParallelAnalyses}
\title{JAMPred step 1 parallel analyses of SNP blocks}
\usage{
JAMPred_Step1ParallelAnalyses(marginal.betas = NULL, n.training = NULL,
  marginal.logor.ses = NULL, n.cases.training = NULL, ref.geno = NULL,
  total.snps.genome.wide = NULL, n.cores = 2, n.mil = 0.2,
  beta.binom.b.lambda = 1, beta.binom.a = 1, snps.blocks = NULL,
  initial.block.size = 100, initial.snps.blocks = NULL, seed = NULL,
  thinning.factor = round(1000/n.mil))
}
\arguments{
\item{marginal.betas}{Vector of named one-at-a-time SNP effects. NB: This must be a named vector; it is where.
the SNP names are derived from. These could be from linear regressions, or log-Odds Ratios. 
NB: If providing log-ORs, you must also provide marginal.logor.ses and n.cases.training so
that a linear transformaiton can be applied.}

\item{n.training}{The sample size in which the marginal.betas were calculated.}

\item{marginal.logor.ses}{IF marginal log-ORs were provided, please provide a named vector of standard errors here.
These are necessary to convert them to a linear scale. Do not set if providing linear regression estimates.}

\item{n.cases.training}{IF marginal log-ORs were provided, please provide the number of cases in the sample in
which they were calculated}

\item{ref.geno}{Reference genotype matrix which will be used to calculate SNP-SNP correlations. Individual's genotype 
must be coded as a numeric risk allele count 0/1/2. Non-integer values reflecting imputation uncertaimnty may be given. 
NB: The risk allele coding MUST correspond to that used in marginal.betas and column names must match the names of
marginal.betas elements.}

\item{total.snps.genome.wide}{Total number of SNPs being analysed across the genome. Used to determine the level of
prior sparsity.}

\item{n.cores}{Number of CPU cores to use. Default is 2.}

\item{n.mil}{Number of million iterations to run. In the paper we found 0.2 was sufficient to reach adequeate
convergence.}

\item{beta.binom.b.lambda}{Spartisty tuning parameter. Suggest a range of values is tried, e.g. 0.01, 0.1, 1, 10.
See the paper for more detail. Default is 1.}

\item{beta.binom.a}{First hyper-parameter of the beta-binomial prior on model sparsity. See the paper for more detail. Default 1 can usually
be used.}

\item{snps.blocks}{OPTIONAL: A partitioning of the SNPs into blocks, created using the function 
\code{\link[R2BGLiMS]{JAMPred_SplitIntoPositiveDefiniteBlocks}}. This is automatically generated 
if not provided.}

\item{initial.block.size}{Block size for the initial partitioning. Default is 100. This is ignored when initial.snps.blocks is given.}

\item{initial.snps.blocks}{A user specficied initial partitioning of the SNPs. This must be a list, each element of which is a vector of SNP
names that correspond to the columns of ref.geno. This could be defined, for example, according to LD information.}

\item{seed}{An integer specifying the RJMCMC seed. If not set a random number will be used every time this is run.}

\item{thinning.factor}{Determines every ith iteration to use from the RJMCMC sample. The more thinning that is applied
the quicker the analysis will be, for the same number of iterations. Leaving at the default should be fine.}
}
\value{
A JAMPred step 1 results object (which is a list of the results and various meta-information).
}
\description{
This function runs the first step of JAMPred - analyses of many SNP blocks independently in parallel.
}
\author{
Paul Newcombe
}
