% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelSizeBayesFactors.R
\name{ModelSizeBayesFactors}
\alias{ModelSizeBayesFactors}
\title{Infer credible sets of predictors}
\usage{
ModelSizeBayesFactors(results)
}
\arguments{
\item{results}{A \code{\link[R2BGLiMS]{R2BGLiMS_Results-class}} object frum running the \code{\link[R2BGLiMS]{R2BGLiMS}} command.}
}
\value{
List of dimension Bayes Factors
}
\description{
Infers Bayes factors for the different possible model sizes from a Reversible Jump Results object.
}
\author{
Tokhir Dadaev and Paul Newcombe
}
