% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NormInvGamPosteriorSample.R
\name{NormInvGamPosteriorSample}
\alias{NormInvGamPosteriorSample}
\title{Generate conjugate posterior sample of coefficients for a particular linear model.}
\usage{
NormInvGamPosteriorSample(data = NULL, outcome.var = NULL,
  confounders = NULL, model = NULL, tau = NULL, n.samples = 1000,
  xTx = NULL, z = NULL, sigma2_invGamma_a = NULL,
  sigma2_invGamma_b = NULL)
}
\arguments{
\item{data}{Matrix or dataframe containing the data to analyse.
Rows are indiviuals, and columns contain the variables and outcome.
If modelling summary statistics specify xTx instead (see below).}

\item{outcome.var}{Name of outcome variable in data. For survival data see times.var below.
If modelling summary statistics specify z instead (see below).}

\item{confounders}{Optional vector of confounders to fix in the model at all times, i.e. exclude from model selection.}

\item{model}{Vector of covariate names to include in the model. Do not include confounders here - they
should be specified with the confounders argument.}

\item{tau}{Value to use for the g-prior sparsity parameter (tau*sigma^2 parameterisation).}

\item{n.samples}{Number of posterior samples to draw.}

\item{xTx}{SUMMARY STATISTICS ONLY: List containing each block's external plug-in estimate for X'X.}

\item{z}{SUMMARY STATISTICS ONLY: t(X)*y vector, calculated from the summary statistics.}

\item{sigma2_invGamma_a}{Inverse-Gamma parameter a for the residual
variance. Not specifying means the value in default.arguments is used (type "data(DefaultArguments)").}

\item{sigma2_invGamma_b}{Inverse-Gamma parameter b for the residual
variance. Not specifying means the value in default.arguments is used (type "data(DefaultArguments)").}
}
\value{
The posterior sample as a matrix.
Rows are different posterior samples, and columns are parameters.
}
\description{
This function can be used to generate a posterior sample of effects under
the Normal-Inverse-Gamma conjugate model, for a particular combination of
covariates (i.e. conditional on a fixed model). For use with JAM specify
xTx and z instead of data and outcome.var.
Note to self: adapted from
Michael Jordan's lecture notes "Bayes Factors, g-priors, and Model
Selection for Regression". Conugate expression is given under the 1/sigma prior.
For the InverseGamma(a, b) prior as below, simply add a and b to the respective
hyperparameters in equation (6) of Michael Jordan's notes.
}
\author{
Paul Newcombe
}

