% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassMethods.R
\docType{methods}
\name{R2BGLiMS_Results-methods}
\alias{R2BGLiMS_Results-methods}
\alias{show,}
\alias{R2BGLiMS_Results-method}
\title{'show' method for \code{\link[R2BGLiMS]{R2BGLiMS_Results-class}} objects}
\usage{
\S4method{show}{R2BGLiMS_Results}(object)
}
\value{
Prints a summary of the \code{\link[R2BGLiMS]{R2BGLiMS_Results-class}} object.
}
\description{
Prints a summary of the \code{\link[R2BGLiMS]{R2BGLiMS_Results-class}} object.
}
\author{
Paul J. Newcombe \email{paul.newcombe@mrc-bsu.cam.ac.uk}
}
