% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StrongestPairwiseCorrelation.R
\name{StrongestPairwiseCorrelation}
\alias{StrongestPairwiseCorrelation}
\title{Strongest Pairwise Correlation}
\usage{
StrongestPairwiseCorrelation(data = NULL, covariates = NULL)
}
\arguments{
\item{data}{Data matrix (or data frame), within which pairwise correlations will be calculated.}

\item{covariates}{Character vector indicating which columns of data should be considered.}
}
\value{
A character vector containing the names of the covariates between which the strongest
pairwise correlation was observed.
}
\description{
Returns list of covariates between which there are the strongest pairwise correlations
}
\author{
Paul Newcombe
}
