% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AutocorrelationPlot.R
\name{AutocorrelationPlot}
\alias{AutocorrelationPlot}
\title{Autocorrelation plot for model selection. Model inclusion of each predictor is
represented across all iterations as a black and white "heatmap".}
\usage{
AutocorrelationPlot(results, plot.title = NULL, vars.to.include = NULL,
  file = NULL, include.var.names = TRUE, var.dictionary = NULL,
  include.post.probs = TRUE, cex.x.axis.ticks = 1, cex.y.axis.ticks = 1,
  cex.y.axis.labels = 1, mar.for.varnames = 8)
}
\arguments{
\item{results}{A \code{\link[R2BGLiMS]{R2BGLiMS_Results-class}} object frum running the \code{\link[R2BGLiMS]{R2BGLiMS}} command.}

\item{plot.title}{Optionally specify a title for the plot.}

\item{file}{Path to a PDF to print plot to.}

\item{include.var.names}{Include variable names perpendicularly below the x-axis. Turn off if it looks too crowded.}

\item{include.post.probs}{Include a barplot of posterior probabilities at the top.}

\item{cex.x.axis.ticks}{Control the size of tick marks for the covariates along the X-axis. Decrease if there are many.}

\item{cex.y.axis.ticks}{Control the size of tick marks for the Y-axis.}

\item{cex.y.axis.labels}{Control the size of the y-labels. Decrease if these are not fitting on.}

\item{mar.for.varnames}{Margin size for the bottom axis. Default is 8 so there is room for perpendicular variable names.
Decrease if not using variable names.}
}
\value{
NA
}
\description{
Generates an autocorrelation plot from a Java MCMC results class object
}
\examples{
library(R2BGLiMS)

### --- Logistic regression with two model space prior components
# Load results from the biopsy logistic regression analysis (see R2MHRJ examples)
data(biopsyResults)
AutocorrelationPlot(biopsyResults)
}
\author{
Paul Newcombe
}

