% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/JAM.R
\name{JAM}
\alias{JAM}
\title{JAM (Joint Analysis of Marginal statistics)}
\usage{
JAM(marginal.betas = NULL, X.ref = NULL)
}
\arguments{
\item{marginal.betas}{Vector of marginal effect estimates to re-analyse with JAM under multivariate models.}

\item{X.ref}{Reference genotype matrix used by JAM to impute the SNP-SNP correlations. If multiple regions are to be
analysed this should be a list containing reference genotype matrices for each region. Individual's genotype
must be coded as a numeric risk allele count 0/1/2. Non-integer values reflecting imputation may be given.
NB: The risk allele coding MUST correspond to that used in marginal.betas. These matrices must each be positive definite and
the column names must correspond to the names of the marginal.betas vector.}

\item{full.mcmc.sampling}{By default JAM only samples models, since the parameters can be analytically integrated out.
If you would like to force JAM to perform full Reversible Jump MCMC of models and parameters (effects and residual), then
set this to TRUE. The posterior summaries can be seen using \code{\link{PrettyResultsTable}}. Note that this option is
forced to false if inference via model enumeration is requested by setting enumerate.up.to.dim>0.}

\item{n.cases}{If the marginal.betas contain log-Odds Ratios, please specify the number of cases
with this option, so that JAM can calculate the case proportion in order to invoke an approximate
transformation between the linear and logistic scales.}
}
\value{
A Reversible Jump results object is returned. This is a list of two elements: "args" which records various modelling
arguments used in the analysis, and "results" - a matrix containing all saved posterior samples from the analysis. Columns
correspond to parameters and each row contains values from a particular itertation, with 0 indicating exclusion from the model.

The function \code{\link{PrettyResultsTable}} can be used to print summary posterior results for all parameters. Other functions
for summarising results are listed under "see also".
}
\description{
Wrapper for JAM (Joint Analysis of Marginal statistics). See the Vignette.
}
\examples{
library(R2BGLiMS)

### --- Linear regression with full MCMC of all parameters
set.seed(1)
# Simulate data
X <- matrix(rnorm(1000, 0, 1), 100,10) # n=100, 10 covariates
confounder <- rnorm(100, 0, 1) 
b <- c(1, rep(0,9), 4) # Only the first covariate has an effect
y <- rnorm(100, cbind(X,confounder)\%*\%b,1)
data <- data.frame(cbind(X,confounder,y))
# Run model treating prior variance of the effects as unknown
# Poisson model space prior on number of effects
poisson.model.space.prior=list( list("Rate"=0.1, "Variables"=paste("V",c(1:10),sep="")) ) 
lm.hierarchical.prior.results <- R2BGLiMS(
  likelihood="Gaussian", data=data, outcome.var="y", confounders="confounder",
  model.space.priors = poisson.model.space.prior
)
PrettyResultsTable(lm.hierarchical.prior.results)
# A numeric representation of the results is stored in the slot 'posterior.summary.table':
lm.hierarchical.prior.results@posterior.summary.table
# Re-analyse using fixed priors
# (not enough covariates to estimate effect variance)
beta.priors <- data.frame(
  cbind(rep(0,11),rep(10,11)),
  row.names=c("confounder",paste("V",c(1:10),sep="")) )
lm.fixed.prior.results <- R2BGLiMS(
  likelihood="Gaussian", data=data, outcome.var="y", confounders="confounder",
  model.space.priors = poisson.model.space.prior,
  beta.priors=beta.priors)
PrettyResultsTable(lm.fixed.prior.results) # Mixing seems a little poorer
# Re-analyse using the conjugate Gaussian model
# Only models are sampled: effects are analytically integrated out
lm.conjugate.results <- R2BGLiMS(
  likelihood="GaussianConj", data=data, outcome.var="y",confounders="confounder",
  model.space.priors=poisson.model.space.prior,
  enumerate.up.to.dim=0,
  tau=100, # Tau recommended to be maximum of P^2 and N
  g.prior=T)
PrettyResultsTable(lm.conjugate.results)
TopModels(lm.conjugate.results)
# Obtain posterior sample of parameter V1
posterior.sample <- NormInvGamPosteriorSample(
  data=data, outcome.var ="y", confounders="confounder",
  model=c("V1"),tau=100)
# Median and credible interval
round(quantile(posterior.sample[,"V1"],c(0.5, 0.025, 0.975)),2)
# Clearly excludes 0

### --- Logistic regression with two model space prior partitions
# Example dataset from MASS
utils::data(biopsy, package = "MASS")
# Normalise predictors in order to use (default) common unknown priors on their effect sizes
for (v in paste("V",c(1:9),sep="")) {
  biopsy[,v] <- (biopsy[,v]-mean(biopsy[,v],na.rm=T))/sd(biopsy[,v],na.rm=T)
}
biopsyResults <- R2BGLiMS( # Takes a few minutes to run
  likelihood="Logistic", data=biopsy, outcome.var="class",
  model.space.priors=list(
    list("Rate"=0.5, "Variables"=paste("V",c(1:3),sep="")),
    list("Rate"=0.1, "Variables"=paste("V",c(4:9),sep=""))
  )
)
PrettyResultsTable(biopsyResults)
TopModels(biopsyResults)

### --- Survival analysis with priors specified on effect sizes
# Example dataset from MASS
utils::data(VA, package = "MASS")
# Generate priors on log-effect sizes
predictors <- c("treat","age","Karn","diag.time","prior")
beta.priors <- data.frame(
  cbind(rep(0,length(predictors)),rep(10,length(predictors))),
  row.names=predictors)
# Analysis using the Weibull likelihood
# Notice how survival analysis requires additionally flagging a "times.var"
va.results.weibull <- R2BGLiMS(
  likelihood="Weibull",
  data=VA,
  outcome.var="status",
  times.var="stime",
  model.space.priors=list(list("a"=1,"b"=1,"Variables"=predictors)),
  beta.prior=beta.priors
)
PrettyResultsTable(va.results.weibull)
TopModels(va.results.weibull)
# Analysis using the Cox likelihood
va.results.cox <- R2BGLiMS(
  likelihood="Cox",
  data=VA,
  outcome.var="status",
  times.var="stime",
  model.space.priors=list(list("a"=1,"b"=1,"Variables"=predictors)),
  beta.prior=beta.priors
)
PrettyResultsTable(va.results.cox)
TopModels(va.results.cox)
}
\author{
Paul Newcombe
}
\seealso{
Summary results are stored in the slot posterior.summary.table. See also
\code{\link{PrettyResultsTable}} and \code{\link{ManhattanPlot}}. For posterior model space
summaries see \code{\link{TopModels}}. For convergence checks
plots see \code{\link{ChainPlots}} and \code{\link{AutocorrelationPlot}}.
}

