% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JAM_PointEstimates.R, R/JAM_PosteriorSampler.R
\name{JAM_PointEstimates}
\alias{JAM_PointEstimates}
\alias{JAM_PosteriorSampler}
\title{JAM (Joint Analysis of Marginal statistics) multivariate point estimator.}
\usage{
JAM_PointEstimates(
  marginal.betas = NULL,
  X.ref = NULL,
  cor.ref = NULL,
  mafs.ref = NULL,
  n = NULL,
  just.get.z = FALSE,
  mafs.if.independent = NULL
)

JAM_PosteriorSampler(marginal.betas = NULL, X.ref = NULL)
}
\arguments{
\item{marginal.betas}{Vector of (named) marginal effect estimates to re-analyse with JAM under multivariate models. 
For multi-ethnic "mJAM" please provide a list of vectors, each element of which is a vector of marginal effects
for a specific ethnicity over the same variants.}

\item{X.ref}{Reference genotype matrix used by JAM to impute the SNP-SNP correlations. If multiple regions are to be 
analysed this should be a list containing reference genotype matrices for each region. Individual's genotype
must be coded as a numeric risk allele count 0/1/2. Non-integer values reflecting imputation may be given.
NB: The risk allele coding MUST correspond to that used in marginal.betas. These matrices must each be positive definite and
the column names must correspond to the names of the marginal.betas vector.}

\item{cor.ref}{Alternatively to a reference genotype matrix, a reference correlation matrix AND mafs may be supplied to JAM.
NB: The risk allele coding MUST correspond to that used in marginal.betas. These matrices must each be positive definite and
the column and row names must correspond to the names of the marginal.betas vector.}

\item{mafs.ref}{Alternatively to a reference genotype matrix, a reference correlation matrix AND mafs may be supplied to JAM.
NB: The risk allele coding MUST correspond to that used in marginal.betas. This must be a named vector with names correspond 
to the names of the marginal.betas vector.}

\item{n}{The size of the dataset in which the summary statistics (marginal.betas) were calculated}

\item{just.get.z}{Return the estimated X'y outcome used in JAM (without actually calculating the corresponding point estimates). This is mainly
for internal use.}

\item{mafs.if.independent}{If the SNPs are independent then a reference genotype matrix is not required.
However, it is still necessary to provide SNP MAFs here as a named vector. Doing so will lead to X.ref being 
ignored and the SNPs to be modelled as if they are independent. Note that this option does not work with 
enumeration.}
}
\value{
A vector of multivariate point estimates.

A vector of multivariate point estimates.
}
\description{
Constructs a point estimate of multivariate SNP effects, based on the JAM estimator, from a vetorc of marginal
effects and a reference genotype matrix.

Constructs a point estimate of multivariate SNP effects, based on the JAM estimator, from a vetorc of marginal
effects and a reference genotype matrix.
}
\seealso{
See also \code{\link{JAM}}.

See also \code{\link{JAM}}.
}
\author{
Paul Newcombe

Paul Newcombe
}
