% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ManhattanPlot.R
\name{ManhattanPlot}
\alias{ManhattanPlot}
\title{Manhattan Plot}
\usage{
ManhattanPlot(results = NULL, plot.quantity = "PosteriorProbability",
  results.vector = NULL, vars.to.include = NULL, var.dictionary = NULL,
  plot.title = "Manhattan Plot", y.max = NULL, add.bonferroni = FALSE,
  top.hits = NULL, X.mat = NULL, point.cols = NULL, point.labs = NULL,
  suppress.x.ticks = FALSE)
}
\arguments{
\item{results}{Reversible Jump results object from running \code{\link{R2BGLiMS}}.}

\item{plot.quantity}{Can be "PosteriorProbability" (default), "BayesFactor" or "pvalue". The
latter is implemented to allow easy comparison between analysis frameworks using the same plot style,
and only works if a vector of p-values is supplied.}

\item{results.vector}{Alternatively can simply pass a named vector of probabilties}

\item{vars.to.include}{Optional character vector specifying a subset of variables
to restrict output to.}

\item{var.dictionary}{Alternative covariate names. This must be a character vector,
named wtih the current covariate names. This can consist of expression functions, e.g. to
achieve italics or superscripts when passed to a plot function.}

\item{plot.title}{Optional character string to name the plot with}

\item{y.max}{If plotting Bayes Factors, this is an optional upper limit for the y-axis (e.g.
to make comparison with a different set of results easier on the eye). Note that if p-values are
being plotted, then this can be used to supply the maximum -log10(p-value) on the y-axis.}

\item{add.bonferroni}{If plotting p-values this option draws a dashed red line at the Bonferroni
threshold. (default is TRUE).}

\item{top.hits}{If a list of "top hits" are provided here, along with the original X matrix below,
then the points are coloured according to their pairwise correlation with the "top hits" (default NULL)}

\item{X.mat}{Must be provided if top.hits are provided above - used to calculate pairwise correlations.}

\item{point.cols}{Optional vector of colours to use for the points (as many and in the same
order as the covariates)}

\item{point.labs}{Optional vector of labels to use for the points (as many and in the same
order as the covariates)}

\item{cex.axis}{Optional size setting for the variable names relative to the x-axis. May want to
adjust if variable names are not legible. (default is 0.5)}

\item{supress.x.ticks}{Set to TRUE to avoid adding x tick mark labels (e.g. if there are many many
variables)}
}
\description{
Produces a Manhattan plot of marginal predictor evidence from a Reversible Jump Results object.
}
\examples{
library(R2BGLiMS)

### --- Logistic regression with two model space prior components
# Load results from the biopsy logistic regression analysis (see R2MHRJ examples)
data(biopsyResults)
ManhattanPlot(biopsyResults)
}
\author{
Paul Newcombe
}

