% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/R2BGLiMS.R
\name{R2BGLiMS}
\alias{R2BGLiMS}
\title{Call BGLiMS from R}
\usage{
R2BGLiMS(likelihood = NULL, data = NULL, outcome.var = NULL,
  times.var = NULL, confounders = NULL, model.selection = TRUE,
  model.space.priors = NULL, beta.priors = NULL, g.prior = FALSE,
  model.tau = FALSE, tau = NULL, enumerate.up.to.dim = 0, X.ref = NULL,
  marginal.betas = NULL, n = NULL, max.fpr = 1, min.tpr = 0,
  n.mil = 1, seed = 1, do.chain.plot = FALSE, results.path = NULL,
  results.label = "R2BGLiMS", extra.arguments = NULL,
  initial.model = NULL, max.model.dim = -1, debug.path = NULL)
}
\arguments{
\item{likelihood}{Type of model to fit. Current options are
"Logistic" (for binary data),
"Weibull" (for survival data),
"Cox" (for survival data),
"RocAUC" (to optimise ROC AUC),
"Gaussian" (for linear regression),
"GaussianConj" (linear regression exploiting conjugate results),
"JAM" (for conjugate linear regression using summary statistics, integrating out parameters)
and "JAM_MCMC" (for linear regression using summary statistics, with full MCMC of all parameters).}

\item{data}{Matrix or dataframe containing the data to analyse.
Rows are indiviuals, and columns contain the variables and outcome.
If modelling summary statistics specify X.ref, marginal.betas, and n instead (see below).}

\item{outcome.var}{Name of outcome variable in data. For survival data see times.var below.
If modelling summary statistics with JAM this can be left null but you must specify X.ref, marginal.beats and n instead (see below).}

\item{times.var}{SURVIVAL DATA ONLY Name of column in data which contains the event times.}

\item{confounders}{Optional vector of confounders to fix in the model at all times, i.e. exclude from model selection.}

\item{model.selection}{Whether to use model selection (default is TRUE). NB: Even if set to FALSE, please provide a
dummy model.space.priors argument (see below). This will not be used mathematically, but the package requires taking the list of variable
names from the "Variables" element therein.}

\item{model.space.priors}{Must be specified if model.selection is set to TRUE.
Two options are available. 1) A fixed prior is placed on the proportion of causal
covariates, and all models with the same number of covariates is equally likely. This
is effectively a Poisson prior over the different possible model sizes. A list must
be supplied for `model.space.priors` with an element "Rate", specifying the prior
proportion of causal covariates, and an element "Variables" containing the list of covariates
included in the model search. 2) The prior proportion of causal covariates
is treated as unknown and given a Beta(a, b) hyper-prior, in which case
elements "a" and "b" must be included in the `model.space.priors` list rather
than "Rate". Higher values of "b" relative to "a" will encourage sparsity.
NOTE: It is easy to specify different model space priors for different collections of
covariates by providing a list of lists, each element of which is itself a model.space.prior
list asm described above for a particular subset of the covariates.}

\item{beta.priors}{There are three options:
1) Leave as null, in which case N(0, 1e6) priors are placed on the effects
of confounders and any predictors included in model selection are assigned a
common normal prior with unknown variance.
2) Provide fixed priors for the confounders only.
3) Provide fixed priors for all covariates.
beta.priors (a matrix/data.frame) can be used to provide fixed priors; rows must be named with the corresponding
variable names in the data, and include Guassian prior means and sds in the first and
second columns respectively.}

\item{g.prior}{CONJUGATE LINEAR REGRESSION ONLY: Gaussian conjugate models ONLY: Whether to use a g-prior for the beta's - i.e. a multivariate normal
with correlation structure proportional to sigma^2*X'X^-1 or to use independence priors (default = FALSE).}

\item{model.tau}{CONJUGATE LINEAR REGRESSION ONLY: Gaussian conjugate models ONLY: Whether to model tau or not (default FALSE). If set to true,
then a Zellner-Siow prior is used, centred on the value provide by tau. The value provided in tau is also used as the
initial value.}

\item{tau}{CONJUGATE LINEAR REGRESSION ONLY: Value to use for sparsity parameter tau (under the tau*sigma^2 parameterisation).
A recommended default is max(n, P^2) where n is the number of individuals, and P the number of predictors.
If modelling this parameter, this value is used to center the Zellner-Siow prior
and as an intial value.}

\item{enumerate.up.to.dim}{CONJUGATE LINEAR REGRESSION ONLY: Whether to calculate posterior scores
for every possible model of dimension up to the integer specified (as an alternative to RJMCMC).
Currenly maximum allowed dimension is 5.
Leaving at 0 means this is not carried out.}

\item{X.ref}{JAM ONLY: Reference genotype matrix with which to impute the SNP-SNP correlations. If multiple regions are to be
analysed this should be a list containing reference genotype matrices for each region. Individual's genotype
must be coded as a numeric risk allele count 0/1/2. Non-integer values reflecting imputation may be given.
NB: The risk allele coding MUST correspond to that used in marginal.betas. These matrices must each be positive definite and
the column names must correspond to the names of the marginal.betas vector.}

\item{marginal.betas}{JAM ONLY: Vector of marginal effect estimates.}

\item{n}{JAM ONLY: Sample size which marginal effects came from.}

\item{max.fpr}{ROC AUC ONLY: Maximum acceptable false positive rate (or x-axis value) to optimise a truncated ROC AUC.}

\item{min.tpr}{ROC AUC ONLY: Minimum acceptable true positive rate, i.e. sensitivity (or y-axis value) to optimise a truncated ROC AUC.}

\item{n.mil}{Number of million iterations to run (default is 1)}

\item{seed}{Which random number seed to use in the RJMCMC sampler.}

\item{do.chain.plot}{Whether to produce a PDF containing chain plots. Default FALSE.}

\item{results.path}{Optional path if wish to to save algorithm output.}

\item{results.label}{Optional label for algorithm output files (if you have specified results.path).}

\item{extra.arguments}{A named list of any additional arguments for BGLiMS. Type "data(DefaultArguments)" and look in the
"default.arguments" list to see the names (which must match) and default values of available extra arguments. Currently these
are not documented - please contact the package author, Paul Newcombe for details.}

\item{initial.model}{Optionally, an initial model can be provided as a vector of 0's and 1's. Default is NULL
and the null model is used. If set to 1, the saturated model is used.}

\item{max.model.dim}{Optional specification of maximum model dimension (default -1 means no maximum is set).}

\item{debug.path}{Optional path to save the data and results files to (rather than as temporary files), for aid in
debugging.}
}
\value{
A Reversible Jump results object is returned. This is a list of two elements: "args" which records various modelling
arguments used in the analysis, and "results" - a matrix containing all saved posterior samples from the analysis. Columns
correspond to parameters and each row contains values from a particular itertation, with 0 indicating exclusion from the model.

The function \code{\link{PrettyResultsTable}} can be used to print summary posterior results for all parameters. Other functions
for summarising results are listed under "see also".
}
\description{
Calls BGLiMS - a Java package for fitting GLMs under Bayesian model selection. NOTE: The predictors to explore with model
selection are specified via the model.space.priors argument - see examples. By Default a common, and unknown, prior SD
is used for all predictors under model selection, which in turn is assigned an Inverse-Gamma hyper-prior.
Fixed normal priors may be user specified for all predictor (and confounder) coefficients via the beta.priors argument.
}
\examples{
library(R2BGLiMS)

### --- Linear regression with full MCMC of all parameters
set.seed(1)
# Simulate data
X <- matrix(rnorm(1000, 0, 1), 100,10) # n=100, 10 covariates
confounder <- rnorm(100, 0, 1) 
b <- c(1, rep(0,9), 4) # Only the first covariate has an effect
y <- rnorm(100, cbind(X,confounder)\%*\%b,1)
data <- data.frame(cbind(X,confounder,y))
# Run model treating prior variance of the effects as unknown
# Poisson model space prior on number of effects
poisson.model.space.prior=list( list("Rate"=0.1, "Variables"=paste("V",c(1:10),sep="")) ) 
lm.hierarchical.prior.results <- R2BGLiMS(
  likelihood="Gaussian", data=data, outcome.var="y", confounders="confounder",
  model.space.priors = poisson.model.space.prior
)
PrettyResultsTable(lm.hierarchical.prior.results, measure="Effect")
# Re-analyse using fixed priors
# (not enough covariates to estimate effect variance)
beta.priors <- data.frame(
  cbind(rep(0,11),rep(10,11)),
  row.names=c("confounder",paste("V",c(1:10),sep="")) )
lm.fixed.prior.results <- R2BGLiMS(
  likelihood="Gaussian", data=data, outcome.var="y", confounders="confounder",
  model.space.priors = poisson.model.space.prior,
  beta.priors=beta.priors)
PrettyResultsTable(lm.fixed.prior.results, measure="Effect") # Mixing seems a little poorer
# Re-analyse using the conjugate Gaussian model
# Only models are sampled: effects are analytically integrated out
lm.conjugate.results <- R2BGLiMS(
  likelihood="GaussianConj", data=data, outcome.var="y",confounders="confounder",
  model.space.priors=poisson.model.space.prior,
  enumerate.up.to.dim=0,
  tau=100, # Tau recommended to be maximum of P^2 and N
  g.prior=T)
PrettyResultsTable(lm.conjugate.results)
TopModels(lm.conjugate.results)
# Obtain posterior sample of parameter V1
posterior.sample <- NormInvGamPosteriorSample(
  data=data, outcome.var ="y", confounders="confounder",
  model=c("V1"),tau=100)
# Median and credible interval
round(quantile(posterior.sample[,"V1"],c(0.5, 0.025, 0.975)),2)

### --- Logistic regression with two model space prior partitions
# Example dataset from MASS
utils::data(biopsy, package = "MASS")
# Normalise predictors in order to use (default) common unknown priors on their effect sizes
for (v in paste("V",c(1:9),sep="")) {
  biopsy[,v] <- (biopsy[,v]-mean(biopsy[,v],na.rm=T))/sd(biopsy[,v],na.rm=T)
}
biopsyResults <- R2BGLiMS(
  likelihood="Logistic", data=biopsy, outcome.var="class",
  model.space.priors=list(
    list("Rate"=0.5, "Variables"=paste("V",c(1:3),sep="")),
    list("Rate"=0.1, "Variables"=paste("V",c(4:9),sep=""))
  )
)
PrettyResultsTable(biopsyResults)
TopModels(biopsyResults)

### --- Survival analysis with priors specified on effect sizes
# Example dataset from MASS
utils::data(VA, package = "MASS")
# Generate priors on log-effect sizes
predictors <- c("treat","age","Karn","diag.time","prior")
beta.priors <- data.frame(
  cbind(rep(0,length(predictors)),rep(10,length(predictors))),
  row.names=predictors)
# Analysis using the Weibull likelihood
# Notice how survival analysis requires additionally flagging a "times.var"
va.results.weibull <- R2BGLiMS(
  likelihood="Weibull",
  data=VA,
  outcome.var="status",
  times.var="stime",
  model.space.priors=list(list("a"=1,"b"=1,"Variables"=predictors)),
  beta.prior=beta.priors
)
PrettyResultsTable(va.results.weibull)
TopModels(va.results.weibull)
# Analysis using the Cox likelihood
va.results.cox <- R2BGLiMS(
  likelihood="Cox",
  data=VA,
  outcome.var="status",
  times.var="stime",
  model.space.priors=list(list("a"=1,"b"=1,"Variables"=predictors)),
  beta.prior=beta.priors
)
PrettyResultsTable(va.results.cox)
TopModels(va.results.cox)
}
\author{
Paul Newcombe
}
\seealso{
For summary results of covariates see \code{\link{ResultsTable}},
\code{\link{PrettyResultsTable}} and \code{\link{ManhattanPlot}}. For posterior model space
summaries see \code{\link{ModelSizes}} and \code{\link{TopModels}}. For convergence check
plots see \code{\link{ChainPlots}} and \code{\link{AutocorrelationPlot}}.
}

