% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TopModels.R
\name{TopModels}
\alias{TopModels}
\title{Table of top models}
\usage{
TopModels(results, n.top.models = 20, covariates.to.include = NULL,
  remove.empty.cols = TRUE)
}
\arguments{
\item{results}{A \code{\link[R2BGLiMS]{R2BGLiMS_Results-class}} object frum running the \code{\link[R2BGLiMS]{R2BGLiMS}} command.}

\item{n.top.models}{number of top models to include in the table (default 20).}

\item{covariates.to.include}{Character vector indicating a subset of covariates to calculate marginal posterior combinations for. NOTE: Not yet implemented for enumeration.}
}
\value{
Ordered table of best models, with corresponding posterior probabilities. If JAM
was run with multiple covariate blocks, a list of model tables is returned.
}
\description{
Generates a table of the top models from a Reversible Jump Results object.
}
\examples{
library(R2BGLiMS)

### --- Logistic regression with two model space prior components
# Load results from the biopsy logistic regression analysis (see R2MHRJ examples)
data(biopsyResults)
TopModels(biopsyResults)
}
\author{
Paul Newcombe
}

